+++ HISTORY.rst
@@ -6,11 +6,26 @@ Changelog
 
 New:
 
+- Implement new `Query Choice` field for collective.easyform which
+  allows you to specify a query to retrieve values from. In order to use,
+  you need to activate the field in the Easy Form Control panel.
+  [vangheem]
+
 - Add `--skip-incomplete` option to `upgrade-sites` script to bypass
   erroring when an profile does not upgrade corrrectly
   [vangheem]
 
+Changes:
+
+- Reorganize `castle.cms.utils` module so split into sub-modules. Imports
+  are all still same.
+  [vangheem]
+
 Fixes:
+
+- Provide patch for https://github.com/celery/celery/pull/4839 until
+  it is fixed in a release
+  [vangheem]
 
 - Skip auto-upgrading `collective.easyform` in `upgrade-sites` script
   because it does not correctly define upgrade steps